Roadmap_Generator/requirements.txt
uvicorn
google-generativeai
fastapi
python-dotenv